# backend/openapi/deadmans.v1.yaml
openapi: 3.0.3
info:
  title: Dead-Mans API
  version: v1
  description: >
    Canonical transport contract for Dead-Mans: HTTP paths under /api and SignalR hubs
    documented in x-signalr. Regenerate frontend contracts after changes.
servers:
  - url: /api
paths:
...
  /game/quiz/questions/available:
    get:
      operationId: getAvailableGameQuizQuestions
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Remaining enabled questions included in the active game
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailableGameQuizQuestionDto'
        '401':
          description: Not authenticated
        '403':
          description: Missing moderator/admin role
  /game/quiz/questions/ask-next:
    post:
      operationId: askNextGameQuizQuestion
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Next question session created immediately when Twitch module is disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskedQuizQuestionDto'
        '202':
          description: Twitch publication preparation persisted; timer has not started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrepareTwitchQuizQuestionResultDto'
        '400':
          description: Authenticated principal is missing required identity claims
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing moderator/admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No active game or no available questions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /game/quiz/questions/{questionId}/ask:
    post:
      operationId: askSpecificGameQuizQuestion
      security:
        - cookieAuth: []
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Selected quiz question session created immediately when Twitch module is disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskedQuizQuestionDto'
        '202':
          description: Twitch publication preparation persisted; timer has not started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrepareTwitchQuizQuestionResultDto'
        '401':
          description: Not authenticated
        '403':
          description: Missing moderator/admin role
        '404':
          description: Question is unavailable
  /game/quiz/current:
    get:
      operationId: getCurrentGameQuizState
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current or most recently closed quiz question session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentGameQuizStateDto'
        '204':
          description: No quiz question session exists
        '401':
          description: Not authenticated
  /game/quiz/question-sessions/{questionSessionId}/submissions:
    post:
      operationId: submitGameQuizAnswer
      security:
        - cookieAuth: []
      parameters:
        - name: questionSessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitGameQuizAnswerRequestDto'
      responses:
        '200':
          description: Answer accepted or idempotently returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameQuizSubmissionReceiptDto'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Question session, option, or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Question session closed or user already selected another option
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /game/quiz/manual-awards/players:
    get:
      operationId: getManualGameQuizAwardPlayers
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Active players and current balances available for manual quiz point adjustments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManualQuizAwardPlayerDto'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing moderator/admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /game/quiz/manual-awards:
    post:
      operationId: awardManualGameQuizPoints
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualQuizAwardRequestDto'
      responses:
        '201':
          description: Manual quiz point adjustment applied or an idempotent prior result returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualQuizAwardSummaryDto'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing moderator/admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No active game or selected player not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Deduction exceeds the available balance or request identifier conflicts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: dm_auth
  schemas:
...
    PrepareTwitchQuizQuestionResultDto:
      type: object
      required: [outcome]
      properties:
        outcome:
          type: string
          enum: [prepared, disabled, notConnected, noActiveGame, noAvailableQuestions, publicationInProgress, pendingOutcome, incompatibleQuestion]
        publication:
          type: object
          allOf:
            - $ref: '#/components/schemas/TwitchQuizPublicationStateDto'
          nullable: true
        errorCode: { type: string, nullable: true }
...
    AskedQuizQuestionDto:
      type: object
      required:
        - questionSessionId
        - gameId
        - askOrder
        - questionId
        - questionCode
        - categoryName
        - text
        - options
        - reward
        - askedAtUtc
        - closesAtUtc
      properties:
        questionSessionId:
          type: string
          format: uuid
        gameId:
          type: string
          format: uuid
        askOrder:
          type: integer
          minimum: 1
        questionId:
          type: string
          format: uuid
        questionCode:
          type: string
        categoryName:
          type: string
        text:
          type: string
        options:
          type: array
          minItems: 2
          maxItems: 10
          items:
            $ref: '#/components/schemas/GameQuizOptionDto'
        reward:
          type: integer
          minimum: 0
        askedAtUtc:
          type: string
          format: date-time
        closesAtUtc:
          type: string
          format: date-time
    GameQuizOptionDto:
      type: object
      required: [optionId, text, displayOrder]
      properties:
        optionId:
          type: string
          format: uuid
        text:
          type: string
        displayOrder:
          type: integer
          minimum: 0
    AvailableGameQuizQuestionDto:
      type: object
      required: [questionId, questionCode, categoryName, text]
      properties:
        questionId:
          type: string
          format: uuid
        questionCode:
          type: string
        categoryName:
          type: string
        text:
          type: string
    SubmitGameQuizAnswerRequestDto:
      type: object
      required: [optionId]
      properties:
        optionId:
          type: string
          format: uuid
    GameQuizSubmissionReceiptDto:
      type: object
      required: [submissionId, questionSessionId, userId, selectedOptionId, submittedAtUtc, isExisting]
      properties:
        submissionId:
          type: string
          format: uuid
        questionSessionId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        selectedOptionId:
          type: string
          format: uuid
        submittedAtUtc:
          type: string
          format: date-time
        isExisting:
          type: boolean
...
    CurrentGameQuizStateDto:
      type: object
      required: [questionSessionId, gameId, askOrder, questionId, questionCode, categoryName, text, options, status, askedAtUtc, closesAtUtc]
      properties:
        questionSessionId:
          type: string
          format: uuid
        gameId:
          type: string
          format: uuid
        askOrder:
          type: integer
          minimum: 1
        questionId:
          type: string
          format: uuid
        questionCode:
          type: string
        categoryName:
          type: string
        text:
          type: string
        options:
          type: array
          minItems: 2
          maxItems: 10
          items:
            $ref: '#/components/schemas/GameQuizOptionDto'
        reward:
          type: integer
          minimum: 0
          nullable: true
          description: Present only after the question timer closes the session.
        status:
          type: string
          enum: [open, closed, skipped]
        askedAtUtc:
          type: string
          format: date-time
        closesAtUtc:
          type: string
          format: date-time
        closedAtUtc:
          type: string
          format: date-time
          nullable: true
        mySelectedOptionId:
          type: string
          format: uuid
          nullable: true
        mySubmittedAtUtc:
          type: string
          format: date-time
          nullable: true
        correctOptionId:
          type: string
          format: uuid
          nullable: true
        myIsCorrect:
          type: boolean
          nullable: true
        myAwardedPoints:
          type: integer
          minimum: 0
          nullable: true
        totalSubmissions:
          type: integer
          minimum: 0
          nullable: true
        optionResults:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/GameQuizOptionResultDto'
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
          nullable: true
          description: Stable machine-readable error code.
          enum:
            - auth.api_client_header_required
            - role_administration.invalid_request
            - role_administration.user_not_found
            - role_administration.permanent_superadmin_protected
            - game_board.cell_not_found
            - game_board.active_team_required
            - game_board.active_team_no_active_game
            - game_board.active_team_not_found
            - game_board.active_team_not_confirmed
            - game_board.active_team_already_played
            - game_board.active_team_has_no_active_members
            - game_board.active_team_round_in_progress
            - game_board.team_played_state_no_active_game
            - game_board.team_played_state_not_found
            - game_board.team_played_state_not_confirmed
            - game_board.team_played_state_round_in_progress
            - game_setup.no_draft
            - game_setup.draft_exists
            - game_setup.invalid_title
            - game_setup.invalid_save_request
            - game_setup.cell_not_found
            - game_setup.cell_media_not_found
            - game_setup.invalid_cell_media_upload
            - game_setup.stale_version
            - game_lifecycle.draft_not_found
            - game_lifecycle.current_already_exists
            - game_lifecycle.active_already_exists
            - game_lifecycle.game_not_ready
            - game_lifecycle.game_not_active
            - game_lifecycle.registration_slots_required
            - game_lifecycle.questions_unavailable
            - game_lifecycle.invalid_team_size_limits
            - game_lifecycle.no_confirmed_teams
            - game_lifecycle.unconfirmed_teams
            - game_lifecycle.pending_invitations
            - game_lifecycle.pending_disband_requests
            - game_lifecycle.invalid_confirmed_team_roster
            - game_lifecycle.operation_failed
            - game_lifecycle.draft_delete_not_allowed
            - game_lifecycle.archive_not_allowed
            - game_lifecycle.game_not_found
            - game_finish.round_in_progress
            - game_finish.stale_version
            - game_finish.warnings_not_acknowledged
            - game_finish.modifier_state_invalid
            - game_finish.invalid_request
            - game_common.unexpected_server_error
            - game_common.too_many_requests
            - game_registration.not_open
            - game_registration.no_slots
            - game_registration.already_on_team
            - game_registration.team_not_found
            - game_registration.team_not_joinable
            - game_registration.team_roster_locked
            - game_registration.not_team_member
            - game_registration.invitation_invalid
            - game_registration.slot_not_found
            - game_registration.slot_not_available
            - game_registration.user_not_found
            - game_registration.pending_invitation
            - game_registration.pending_outgoing_invitation
            - game_registration.team_invite_not_allowed
            - game_registration.team_active_in_game
            - game_registration.team_already_played
            - game_registration.invalid_team_name
            - game_registration.team_name_taken
            - game_registration.team_name_required
            - game_registration.team_not_full
            - game_registration.disband_request_not_owned
            - game_registration.operation_failed
            - game_modifier.game_not_active
            - game_modifier.not_enabled
            - game_modifier.emergency_disabled
            - game_modifier_content_locked
            - game_modifier_revision_stale
            - game_modifier_compatibility_locked
            - game_modifier_archived
            - game_modifier_version_binding_missing
            - game_modifier.conflict_active
            - game_modifier.limit_reached
            - game_modifier.ordering_closed
            - game_modifier.active_team_member
            - game_modifier.insufficient_quiz_points
            - game_modifier.player_not_found
            - game_modifier.activation_not_found
            - game_modifier.activation_cancel_forbidden
            - game_modifier.activation_cancel_invalid_state
            - game_modifier.activation_cancel_reason_required
            - game_modifier.user_not_resolved
            - game_modifier.invalid_request
            - game_modifier_not_found
            - game_round.no_active_game
            - game_round.cell_not_found
            - game_round.cell_not_open
            - game_round.team_not_found
            - game_round.team_not_confirmed
            - game_round.team_has_no_active_members
            - game_round.awaiting_modifiers_required
            - game_round.already_in_progress
            - game_round.invalid_request
            - game_round.not_found
            - game_round.not_in_progress
            - game_round.stale_version
            - game_round.modifier_result_not_found
            - modifier_resolution.duplicate_group
            - modifier_resolution.duplicate_result
            - modifier_resolution.result_set_mismatch
            - modifier_resolution.group_set_mismatch
            - modifier_resolution.group_missing
            - modifier_resolution.group_members_mismatch
            - modifier_resolution.violation_comment_required
            - modifier_resolution.automatic_input_forbidden
            - modifier_resolution.boolean_required
            - modifier_resolution.non_negative_count_required
            - modifier_resolution.unsupported
            - modifier_resolution.missing
            - modifier_calculation.failed
            - behavior.invalid
            - behavior.rule_incompatible
            - formula.unsupported
            - formula.incompatible
            - resolution.invalid
            - round_facts.invalid
            - activation.duplicate
            - resolution.rule_status_required
            - resolution.automatic_required
            - resolution.boolean_required
            - resolution.non_negative_count_required
            - resolution.count_exceeds_resolved_kills
            - resolution.count_exceeds_activation_limit
            - resolution.per_activation_required
            - game_question.invalid_request
            - game_question.duplicate_code
            - game_question.not_found
            - game_question.category_not_found
            - game_question.category_not_empty
            - game_question.category_protected
            - game_question.import_invalid_fields
            - game_question.import_duplicate_code_in_file
            - game_question.import_category_unresolved
            - game_question.import_duplicate_code_existing
            - game_quiz.no_active_game
            - game_quiz.no_available_questions
            - game_quiz.answer_player_not_found
            - game_quiz.question_session_not_found
            - game_quiz.question_session_closed
            - game_quiz.already_answered
            - game_quiz.option_not_found
            - game_quiz.manual_award_player_not_found
            - game_quiz.manual_award_invalid_points
            - game_quiz.manual_award_invalid_operation
            - game_quiz.manual_award_invalid_reason
            - game_quiz.manual_award_insufficient_points
            - game_quiz.manual_award_duplicate_request_conflict
            - twitch_quiz.disabled
            - twitch_quiz.not_connected
            - twitch_quiz.publication_in_progress
            - twitch_quiz.pending_outcome
            - twitch_quiz.incompatible_question
            - twitch_quiz.question_message_too_long
            - twitch_quiz.options_message_too_long
            - twitch_quiz.result_message_too_long
        requestId:
          type: string
          nullable: true
          description: Server request correlation identifier for diagnostics.
...